move processOTLPRequest to router
