+++ route/route.go
@@ -32,6 +32,8 @@ import (
 	// grpc/gzip compressor, auto registers on import
 	_ "google.golang.org/grpc/encoding/gzip"
 
+	huskyotlp "github.com/honeycombio/husky/otlp"
+
 	"github.com/honeycombio/refinery/collect"
 	"github.com/honeycombio/refinery/config"
 	"github.com/honeycombio/refinery/internal/health"
@@ -497,6 +499,41 @@ func (r *Router) batch(w http.ResponseWriter, req *http.Request) {
 		return
 	}
 	w.Write(response)
+}
+
+func (router *Router) processOTLPRequest(
+	ctx context.Context,
+	batches []huskyotlp.Batch,
+	apiKey string) error {
+
+	var requestID types.RequestIDContextKey
+	apiHost := router.Config.GetHoneycombAPI()
+
+	// get environment name - will be empty for legacy keys
+	environment, err := router.getEnvironmentName(apiKey)
+	if err != nil {
+		return nil
+	}
+
+	for _, batch := range batches {
+		for _, ev := range batch.Events {
+			event := &types.Event{
+				Context:     ctx,
+				APIHost:     apiHost,
+				APIKey:      apiKey,
+				Dataset:     batch.Dataset,
+				Environment: environment,
+				SampleRate:  uint(ev.SampleRate),
+				Timestamp:   ev.Timestamp,
+				Data:        ev.Attributes,
+			}
+			if err = router.processEvent(event, requestID); err != nil {
+				router.Logger.Error().Logf("Error processing event: " + err.Error())
+			}
+		}
+	}
+
+	return nil
 }
 
 func (r *Router) processEvent(ev *types.Event, reqID interface{}) error {
